# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This service takes facility and biomass location information, along with various configuration parameters, and determines the opimal cluster locations for harvesting.  See https://github.com/ucdavis/cecdss-backend for more details.",
    "version": "1.0.0",
    "title": "CECDSS-Backend API",
    "contact": {
      "email": "srkirkland@ucdavis.edu"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/ucdavis/cecdss-backend/blob/master/LICENSE"
    }
  },
  "host": "virtserver.swaggerhub.com",
  "basePath": "/ucdavis/CECDSS-Backend/1.0.0",
  "schemes": ["https"],
  "paths": {
...
    "/process": {
      "post": {
        "tags": ["main"],
        "summary": "process facility biomass for single year",
        "description": "Used for annual biomass supply location and optimization\n",
        "operationId": "process",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "params",
            "description": "Main params",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RequestParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/YearlyResult"
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      }
    },
...
  },
  "definitions": {
    "RequestParams": {
      "type": "object",
      "required": [
        "annualGeneration",
        "biomassTarget",
        "carbonCreditPrice",
        "cashFlow",
        "clusterIds",
        "costOfEquity",
        "dieselFuelPrice",
        "energyEconomyRatio",
        "errorIds",
        "facilityLat",
        "facilityLng",
        "firstYear",
        "generalInflation",
        "includeCarbonCredit",
        "lat",
        "lng",
        "moistureContent",
        "radius",
        "system",
        "teaModel",
        "treatmentid",
        "year"
      ],
      "properties": {
        "facilityLat": {
          "type": "number",
          "title": "RequestParams.facilityLat"
        },
        "facilityLng": {
          "type": "number",
          "title": "RequestParams.facilityLng"
        },
        "lat": {
          "type": "number",
          "title": "RequestParams.lat"
        },
        "lng": {
          "type": "number",
          "title": "RequestParams.lng"
        },
        "system": {
          "type": "string",
          "title": "RequestParams.system"
        },
        "treatmentid": {
          "type": "number",
          "title": "RequestParams.treatmentid"
        },
        "dieselFuelPrice": {
          "type": "number",
          "title": "RequestParams.dieselFuelPrice"
        },
        "biomassTarget": {
          "type": "number",
          "title": "RequestParams.biomassTarget"
        },
        "firstYear": {
          "type": "number",
          "title": "RequestParams.firstYear"
        },
        "year": {
          "type": "number",
          "title": "RequestParams.year"
        },
        "clusterIds": {
          "type": "array",
          "title": "RequestParams.clusterIds",
          "items": {
            "type": "string",
            "title": "RequestParams.clusterIds.[]"
          }
        },
        "errorIds": {
          "type": "array",
          "title": "RequestParams.errorIds",
          "items": {
            "type": "string",
            "title": "RequestParams.errorIds.[]"
          }
        },
        "radius": {
          "type": "number",
          "title": "RequestParams.radius"
        },
        "teaModel": {
          "type": "string",
          "title": "RequestParams.teaModel"
        },
        "annualGeneration": {
          "type": "number",
          "title": "RequestParams.annualGeneration"
        },
        "moistureContent": {
          "type": "number",
          "title": "RequestParams.moistureContent"
        },
        "cashFlow": {
          "title": "RequestParams.cashFlow"
        },
        "costOfEquity": {
          "type": "number",
          "title": "RequestParams.costOfEquity"
        },
        "generalInflation": {
          "type": "number",
          "title": "RequestParams.generalInflation"
        },
        "carbonCreditPrice": {
          "type": "number",
          "title": "RequestParams.carbonCreditPrice"
        },
        "energyEconomyRatio": {
          "type": "number",
          "title": "RequestParams.energyEconomyRatio"
        },
        "includeCarbonCredit": {
          "type": "boolean",
          "title": "RequestParams.includeCarbonCredit"
        }
      },
      "title": "RequestParams"
    },
...
    "YearlyResult": {
      "type": "object",
      "required": [
        "cashFlow",
        "clusterNumbers",
        "clusters",
        "energyRevenueRequired",
        "energyRevenueRequiredPW",
        "errorClusterNumbers",
        "errorClusters",
        "errorGeoJson",
        "feedstockCostPerTon",
        "fuelCost",
        "geoJson",
        "harvestCostPerDryTon",
        "moveInCostPerDryTon",
        "numberOfClusters",
        "radius",
        "totalArea",
        "totalCoproduct",
        "totalCoproductCost",
        "totalDryCoproduct",
        "totalDryFeedstock",
        "totalFeedstock",
        "totalHarvestCost",
        "totalMoveInCost",
        "totalMoveInDistance",
        "totalTransportationCost",
        "transportationCostPerDryTon",
        "tripGeometries",
        "year"
      ],
      "properties": {
        "tripGeometries": {
          "title": "YearlyResult.tripGeometries"
        },
        "year": {
          "type": "number",
          "title": "YearlyResult.year"
        },
        "lcaResults": {
          "title": "YearlyResult.lcaResults"
        },
        "totalArea": {
          "type": "number",
          "title": "YearlyResult.totalArea"
        },
        "totalFeedstock": {
          "type": "number",
          "title": "YearlyResult.totalFeedstock"
        },
        "totalDryFeedstock": {
          "type": "number",
          "title": "YearlyResult.totalDryFeedstock"
        },
        "totalHarvestCost": {
          "type": "number",
          "title": "YearlyResult.totalHarvestCost"
        },
        "totalCoproduct": {
          "type": "number",
          "title": "YearlyResult.totalCoproduct"
        },
        "totalDryCoproduct": {
          "type": "number",
          "title": "YearlyResult.totalDryCoproduct"
        },
        "totalCoproductCost": {
          "type": "number",
          "title": "YearlyResult.totalCoproductCost"
        },
        "totalMoveInCost": {
          "type": "number",
          "title": "YearlyResult.totalMoveInCost"
        },
        "totalMoveInDistance": {
          "type": "number",
          "title": "YearlyResult.totalMoveInDistance"
        },
        "totalTransportationCost": {
          "type": "number",
          "title": "YearlyResult.totalTransportationCost"
        },
        "harvestCostPerDryTon": {
          "type": "number",
          "title": "YearlyResult.harvestCostPerDryTon"
        },
        "transportationCostPerDryTon": {
          "type": "number",
          "title": "YearlyResult.transportationCostPerDryTon"
        },
        "moveInCostPerDryTon": {
          "type": "number",
          "title": "YearlyResult.moveInCostPerDryTon"
        },
        "feedstockCostPerTon": {
          "type": "number",
          "title": "YearlyResult.feedstockCostPerTon"
        },
        "numberOfClusters": {
          "type": "number",
          "title": "YearlyResult.numberOfClusters"
        },
        "clusterNumbers": {
          "type": "array",
          "title": "YearlyResult.clusterNumbers",
          "items": {
            "type": "string",
            "title": "YearlyResult.clusterNumbers.[]"
          }
        },
        "clusters": {
          "title": "YearlyResult.clusters"
        },
        "errorClusters": {
          "title": "YearlyResult.errorClusters"
        },
        "errorClusterNumbers": {
          "type": "array",
          "title": "YearlyResult.errorClusterNumbers",
          "items": {
            "type": "string",
            "title": "YearlyResult.errorClusterNumbers.[]"
          }
        },
        "radius": {
          "type": "number",
          "title": "YearlyResult.radius"
        },
        "fuelCost": {
          "type": "number",
          "title": "YearlyResult.fuelCost"
        },
        "energyRevenueRequired": {
          "type": "number",
          "title": "YearlyResult.energyRevenueRequired"
        },
        "energyRevenueRequiredPW": {
          "type": "number",
          "title": "YearlyResult.energyRevenueRequiredPW"
        },
        "cashFlow": {
          "title": "YearlyResult.cashFlow"
        },
        "geoJson": {
          "title": "YearlyResult.geoJson"
        },
        "errorGeoJson": {
          "title": "YearlyResult.errorGeoJson"
        }
      },
      "title": "YearlyResult"
    },
...
  }
}